Update evaluation summary with new component
